rtaspi-modular/rtaspi-core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rtaspi-core"
version = "0.2.0"
description = "Core functionality for RTASPI system"
authors = [{name = "RTASPI Team"}]
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",           # Configuration file handling
    "python-dotenv>=1.0.0",  # Environment variable management
    "structlog>=23.1.0",     # Structured logging
    "typing-extensions>=4.5.0", # Enhanced typing support
    "pydantic>=2.0.0",       # Data validation
    "aiohttp>=3.8.0",        # Async HTTP client/server
    "python-json-logger>=2.0.0" # JSON logging format
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pylint>=2.17.0"
]

[tool.setuptools]
packages = ["rtaspi_core"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pylint.messages_control]
disable = [
    "C0111",  # Missing docstring
    "C0103",  # Invalid name
    "C0330",  # Wrong hanging indentation
    "C0326",  # Bad whitespace
]

[tool.pylint.format]
max-line-length = 100
